Rename chat folder to azure-communication-chat; Update swagger README.md to link to swagger; Regenerate with swagger link
