+++ sdk/communication/azure-communication-chat/.gitignore
