+++ sdk/communication/azure-communication-chat/build.gradle
