+++ sdk/communication/azure-communication-chat/proguard-rules.pro
