+++ sdk/communication/azure-communication-chat/src/main/AndroidManifest.xml
