+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/SendReadReceiptRequest.java
