+++ sdk/communication/azure-communication-chat/src/test/java/com/azure/android/communication/chat/MessagesMocker.java
