+++ sdk/communication/azure-communication-chat/swagger/README.md
@@ -20,7 +20,7 @@ autorest --android --use=C:/work/autorest.java
 
 ### Code generation settings
 ``` yaml
-input-file: ./communicationserviceschat.json
+input-file: https://raw.githubusercontent.com/Azure/azure-rest-api-specs/838c5092f11e8ca26e262b1f1099d5c5cdfedc3f/specification/communication/data-plane/Microsoft.CommunicationServicesChat/preview/2020-09-21-preview2/communicationserviceschat.json
 java: true
 output-folder: ..\
 generate-client-as-impl: false
