+++ settings.gradle
@@ -1,4 +1,4 @@
-include ':sdk:communication:chat'
+include ':sdk:communication:azure-communication-chat'
 include ":sdk:core:azure-core"
 include ":sdk:storage:azure-storage-blob"
 include ":sdk:communication:azure-communication-common"
